+++ src/dust3r/patch_embed.py
@@ -52,7 +52,7 @@ class ManyAR_PatchEmbed(PatchEmbed):
 
     def forward(self, img, true_shape):
         B, C, H, W = img.shape
-
+        
         assert (
             H % self.patch_size[0] == 0
         ), f"Input image height ({H}) is not a multiple of patch size ({self.patch_size[0]})."
@@ -77,13 +77,13 @@ class ManyAR_PatchEmbed(PatchEmbed):
         pos = img.new_zeros((B, n_tokens, 2), dtype=torch.int64)
 
         x[is_landscape] = (
-            self.proj(img[is_landscape]).permute(0, 2, 3, 1).flatten(1, 2).float()
+            self.proj(img[is_landscape]).permute(0, 2, 3, 1).flatten(1, 2).to(x.dtype)
         )
         x[is_portrait] = (
             self.proj(img[is_portrait].swapaxes(-1, -2))
             .permute(0, 2, 3, 1)
             .flatten(1, 2)
-            .float()
+            .to(x.dtype)
         )
 
         pos[is_landscape] = self.position_getter(1, H, W, pos.device)
